// src/__generated__/scheduleExecutionsRelay_query.graphql.ts
export type scheduleExecutionsRelay_query$data = {
  readonly scheduleExecutions: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly scriptName: string | null | undefined;
        readonly " $fragmentSpreads": FragmentRefs<"executionFields_execution">;
      };
    }>;
    readonly filteredCount: number;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
  };
  readonly " $fragmentType": "scheduleExecutionsRelay_query";
};
export type scheduleExecutionsRelay_query$key = {
  readonly " $data"?: scheduleExecutionsRelay_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"scheduleExecutionsRelay_query">;
};